+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Avoid duplicate NINA connection entries
 - Eslint now fixes prettier
+- Flatpanel icon state
 
 ## [1.0.4.4] - 2025-03-20
 ### Fixed